requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
flasgger==0.9.7.1
mysql-connector-python==8.2.0
